Begin moving code for language selection out of diagnostic and into shared.
