+++ services/QuillLMS/client/app/bundles/Diagnostic/components/renderForQuestions/cues.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { Cue, CueExplanation } from '../../../Shared/index';
 import { Question } from '../../interfaces/Question';
-import { ENGLISH } from '../../modules/translation/languagePageInfo';
+import { ENGLISH } from '../../../Shared/utils/languageList';
 const arrow = `${process.env.CDN_URL}/images/icons/pointing-arrow.svg`;
 
 interface CuesProps {
